requirements.txt
PySide6
google-generativeai
Pillow
mss
pynput
python-dotenv
pyinstaller
